mark_row_created as common method to insert and update
