+++ core/rs/core/src/local_writes/mod.rs
@@ -3,17 +3,20 @@ use core::mem::ManuallyDrop;
 
 use crate::alloc::string::ToString;
 use crate::c::crsql_ExtData;
+use crate::stmt_cache::reset_cached_stmt;
 use alloc::boxed::Box;
 use alloc::format;
 use alloc::slice;
 use alloc::string::String;
 use alloc::vec::Vec;
-use sqlite::{Context, Value};
+use sqlite::sqlite3;
+use sqlite::{value, Context, ManagedStmt, Value};
 use sqlite_nostd as sqlite;
 use sqlite_nostd::ResultCode;
 
 use crate::tableinfo::{crsql_ensure_table_infos_are_up_to_date, TableInfo};
 
+mod after_insert;
 mod after_update;
 
 pub fn trigger_fn_preamble<F>(
@@ -53,4 +56,50 @@ where
     };
 
     f(table_info, &values, ext_data)
+}
+
+fn step_trigger_stmt(stmt: &ManagedStmt) -> Result<ResultCode, String> {
+    match stmt.step() {
+        Ok(ResultCode::DONE) => {
+            reset_cached_stmt(stmt.stmt)
+                .or_else(|_e| Err("done -- unable to reset cached trigger stmt"))?;
+            Ok(ResultCode::OK)
+        }
+        Ok(code) | Err(code) => {
+            reset_cached_stmt(stmt.stmt)
+                .or_else(|_e| Err("error -- unable to reset cached trigger stmt"))?;
+            Err(format!(
+                "unexpected result code from tigger_stmt.step: {}",
+                code
+            ))
+        }
+    }
+}
+
+fn mark_new_pk_row_created(
+    db: *mut sqlite3,
+    tbl_info: &TableInfo,
+    pks_new: &[*mut value],
+    db_version: i64,
+    seq: i32,
+) -> Result<ResultCode, String> {
+    let mark_locally_created_stmt_ref = tbl_info
+        .get_mark_locally_created_stmt(db)
+        .or_else(|_e| Err("failed to get mark_locally_created_stmt"))?;
+    let mark_locally_created_stmt = mark_locally_created_stmt_ref
+        .as_ref()
+        .ok_or("Failed to deref sentinel stmt")?;
+
+    for (i, pk) in pks_new.iter().enumerate() {
+        mark_locally_created_stmt
+            .bind_value(i as i32 + 1, *pk)
+            .or_else(|_e| Err("failed to bind pks to mark_locally_created_stmt"))?;
+    }
+    mark_locally_created_stmt
+        .bind_int64(pks_new.len() as i32 + 1, db_version)
+        .and_then(|_| mark_locally_created_stmt.bind_int(pks_new.len() as i32 + 2, seq))
+        .and_then(|_| mark_locally_created_stmt.bind_int64(pks_new.len() as i32 + 3, db_version))
+        .and_then(|_| mark_locally_created_stmt.bind_int(pks_new.len() as i32 + 4, seq))
+        .or_else(|_| Err("failed binding to mark_locally_created_stmt"))?;
+    step_trigger_stmt(mark_locally_created_stmt)
 }
